backend/requirements.txt
asgiref==3.9.0
attrs==25.3.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
gunicorn==21.2.0
inflection==0.5.1
itsdangerous==2.2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
